remove front matter, set up overview infrastructure
