+++ app/components/Dataset/_Dataset.scss
@@ -5,11 +5,6 @@
 	height: 800px;
 }
 
-#header .btn, #header .btn-group {
-	margin-right: 10px;
-	margin-bottom: 15px;
-}
-
 #header .btn-group > .btn {
 	margin-bottom: 0;
 	margin-right: 0;
@@ -27,7 +22,10 @@
 #ChartsContainer .toggle-time {
 	float: right;
 }
-
+.toggle-time{
+  margin-top:4%;
+  margin-right:7%;
+}
 #ChartsContainer {
 	margin-top: 20px;
 }
@@ -50,6 +48,158 @@
 }
 
 .c3 svg {
-    font: 12px "Lato", "Helvetica Neue", Helvetica, Arial, sans-serif;
-    line-height: 1.4;
+     line-height: 1.4;
+    font-family: 'Lato', sans-serif;
+}
+
+.toggle-time {
+	float: right;
+}
+
+.homepage{
+  margin:10%
+}
+
+#search-catalog-input::-webkit-input-placeholder {
+    color:    #99b5c3;
+ }
+search-catalog-input:-moz-placeholder {
+    color:    #99b5c3;
+}
+#search-catalog-input::-moz-placeholder {
+    color:    #99b5c3;
+  }
+
+.search-input {
+ font-size: 120%;
+      height: 60px;
+
+    padding-right:2%;
+    margin-right:0.5%;
+}
+
+
+.glyphicon.glyphicon-search {
+    font-size: 28px;
+}
+
+input-mysize{
+  width: 550px;
+}
+
+
+.descriptionPanel{
+  minHeight: '200px';
+  font-size:105%;
+}
+.descriptionPanelHead{
+  margin-top:3%;
+  margin-left:2%;
+  margin-right:2%;
+}
+.descriptionPanelField{
+  font-size:85%;
+}
+
+.dataSetInfo{
+  font-size:110%;
+  margin-top:-0.25%;
+  margin-left:-2%;
+  font-family: 'Lato', sans-serif;
+}
+
+.dataSetInfoDescription{
+  font-size:130%;
+  margin-right:10%;
+  font-family: 'Lato', sans-serif;
+}
+
+.datasetOverview{
+  margin-top: 2%;
+}
+.datasetName{
+}
+
+.datasetLinks{
+  font-size:130%;
+}
+
+.datasetDownLoadButtons{
+  margin-top:1%;
+  margin-left:-2%;
+  font-family: 'Lato', sans-serif;
+}
+
+
+.dataSetTitle{
+   margin-top:6%;
+   font-family: 'Lato', sans-serif;
+}
+
+.chartButtons{
+  font-size:110%;
+  padding:0.50%;
+  border-radius: 0.10 !important;
+  border:0.001;
+  margin-left:0.01%;
+  margin-right:0.01%;
+  margin-bottom:0.01%;
+  font-family: 'Lato', sans-serif;
+}
+
+.chartButtonRow{
+  margin-right:2%;
+  margin-top:2%;
+}
+
+.c3chart{
+  margin-top:3%;
+}
+
+
+.chartFiltersStyle{
+  font-size:115%;
+  padding-right:5%;
+  margin-top:8%;
+}
+
+.filterF{
+  margin-top:2%;
+  margin-bottom:2%;
+  margin-right:3%;
+  font-size:115%;
+  min-width:320px;
+  height:40px;
+}
+h2.chartTitle{
+  margin-top:4%;
+  margin-left:3%;
+}
+
+h3.chartTitle{
+  margin-left:3%;
+  padding-bottom:0.5%;
+}
+
+.chartTypeRadio{
+  font-size:124%;
+  font-family: 'Lato', sans-serif;
+}
+
+.chartTypeRadioTitle{
+  font-size:140%;
+  padding-left:3%;
+  padding-bottom:1.5%;
+  font-family: 'Lato', sans-serif;
+}
+.chartTypeRadioGroup{
+  padding-left:3%;
+    background-color:#d3d3d3;
+  padding-top:3%;
+  font-family: 'Lato', sans-serif;
+  padding-bottom:3%;
+}
+.chartTypeRadioLabel{
+  padding-right:1%;
+  font-family: 'Lato', sans-serif;
 }
