+++ app/components/Navigation/_Navigation.scss
@@ -1,6 +1,5 @@
 @import "toolkit";
 
-
 .myNav .navbar-brand {
   height: 50px;
 }
@@ -14,15 +13,12 @@
   margin: 25px 15px 25px 0;
 }
 
-
-
 .navLogo{
-  font-size:260%;
+  font-size:220%;
   font-family: 'Lato', sans-serif;
-  padding-top:14%;
+  padding-top:12%;
   margin-left:5%;
 }
-
 
 .navTotal{
   width:100%;
@@ -31,9 +27,10 @@
   padding-left:2%;
   padding-right:2%;
 }
+
 .navLi {
   padding-top:2%;
-  font-size:150%;
+  font-size:120%;
   padding-right:2%;
   font-family: 'Lato', sans-serif;
 }
