+++ app/components/_Dataset.scss
@@ -5,10 +5,6 @@
 	height: 800px;
 }
 
-#header .btn, #header .btn-group {
-	margin-bottom: 15px;
-}
-
 #header .btn-group > .btn {
 	margin-bottom: 0;
 	margin-right: 0;
@@ -91,15 +87,6 @@ input-mysize{
   width: 550px;
 }
 
-.tabSelected{
-  font-size:150%;
-}
-
-.chartTabs{
-  margin-top:1.75%;
-  margin-right:1%;
-  margin-left:1%;
-}
 
 .descriptionPanel{
   minHeight: '200px';
@@ -146,7 +133,6 @@ input-mysize{
 
 .dataSetTitle{
    margin-top:6%;
-   margin-right:3%;
    font-family: 'Lato', sans-serif;
 }
