+++ app/scss/bootstrap/pre-customizations.scss
@@ -89,7 +89,7 @@ $icon-font-svg-id:        "glyphicons_halflingsregular" !default;
 //## Define common padding and border radius sizes and more. Values based on 14px text and 1.428 line-height (~20px to start).
 
 $padding-base-vertical:     10px !default;
-$padding-base-horizontal:   15px !default;
+$padding-base-horizontal:   20px !default;
 
 $padding-large-vertical:    18px !default;
 $padding-large-horizontal:  27px !default;
